backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
Pillow
python-dotenv
